chore: Set up release-please
This configures the [release-please action][1] for `generator-eslint`.
As we merge changes using [Conventional Commits formatting][2],
release-please will create and update a pending release PR. That PR will
contain a changelog of any commits merged since the prior release. When
we're ready to do the release, we'll merge that PR, and the workflow in
this PR will update the changelog and publish the release to GitHub and
npm.

Separate from the code in this PR, I had to take a few manual steps:

1. Create a new [npm granular access token][3].
   1. Name it `release-please $package-name`.
   1. Choose a reasonable expiration. I went with 365 days.
   1. Under Packages and Scopes > select "Read and write" access >
      select "Only select packages and scopes" > select the package that
      we're setting up.
   1. The token should not have Organization access.
   1. Generate the token.
1. Copy the generated token as a [new repository secret][4] named
   `NPM_TOKEN`.
1. In the [repository's Actions settings][5], under "Workflow
   permissions", ensure "Allow GitHub Actions to create and approve pull
   requests" is checked.

[1]: https://github.com/google-github-actions/release-please-action
[2]: https://www.conventionalcommits.org/en/v1.0.0/
[3]: https://www.npmjs.com/settings/eslintbot/tokens/granular-access-tokens/new
[4]: https://github.com/eslint/generator-eslint/settings/secrets/actions/new
[5]: https://github.com/eslint/generator-eslint/settings/actions